Read chart location in as arg
If one file path is given as an argument, the chart is modified in
place, leaving a backup file of the original.
Providing no args reads from stdin and to stdout as before.
The arg parser is set up to allow a second path to be specified as an
output path, but this feature is not fully implemented yet.

Issue: #1
Signed-off-by: SyfSchydea <sashapwn@gmail.com>